Creating interfaces for Clients
In order to allow us to mock the clients as we please to be able to
write better tests, we're creating a more abstract layer over the client
structs, having the contracts interfaced instead of directly
implemented.

As mentioned above, this should make the code a bit less coupled as well
as allow us for more thorough testing using Mocks.